Implement utility methods on Chaincode type
Closes #47.

Implements ToChaincode, Bytes, (Un)MarshalText, (Un)MarshalMsg.

This means that Chaincode objects can now be natively used in JSON,
TOML, and MsgPack streams, and for everything else, there are simple
workaround to move to and from byte slices.

I'm not sure _how_ exactly this works--I'd had the impression that
you weren't supposed to be able to define methods on non-struct
types--but it's cool that it does. `make test` passes without issue.