Generate constants from ndau defs; fix makefile
This reads the "chain:" definitions from the ndau repo and uses them to construct the constants to pass to chasm.

Along the way, I discovered that the makefile had a few issues so I fixed them as well.

Also tweaked the wording of the "DO NOT EDIT" lines so that they play nicer with editors that look for that pattern.