Update chaincode handling of nil pointers
In the past, nil pointers were ignored. This caused errors when
otherwise valid structs which contained pointers stopped being
comparable when one, but not both, of them had a nil pointer.

Now, nil pointers are represented in chaincode by the zero value
for that type. It means that chaincode scripts have to watch out
for that case, but the go code gets simpler.